+++ assets/css/partials/_variables.scss
@@ -23,5 +23,8 @@ $pink: #ff80d4;
 $white: #fff;
 $purple: #aa80ff;
 
+$elementDefaultBackground: linear-gradient(136deg, #272f3f 0%, #1d232f 100%);
+
 //->Sizes
-$elementBoxSize: 3.8vw;
+$elementBoxSize: 72px;
+
