+++ components/Slider.vue
@@ -1,67 +1,54 @@
 <template>
-  <div>
-    <div style="width: 30vw; display: flex; justify-content: space-between">
-      <div class="slider-title">Sıcaklık</div>
-      <div style="display: flex; cursor: pointer" @click="onTemperatureChange($store.state.selectedTemperatureType === 'c' ? 25 : 77)">
+  <div class="slider">
+    <div class="w-full flex justify-between mb-4">
+      <div class="slider-title flex items-center">
+        Sıcaklık
+        <toggle-button v-model="active" color="#66ccca" :sync="true" class="ml-3" />
+      </div>
+      <div v-if="active" style="display: flex; cursor: pointer" @click="onTemperatureChange($store.state.unit === 'c' ? 25 : 77)">
         <img src="../assets/icons/icons-yenile-gray.svg" style="margin-right: 5px" />
         <div class="slider-title" style="font-weight: bold; margin-top: 2px">Yenile</div>
       </div>
     </div>
-    <client-only>
-      <div style="display: flex; margin-top: 8px" class="table-slider">
-        <vue-range-slider
-          ref="slider"
-          :value="viewTemperature"
-          :tooltip="'none'"
-          :min="minDegree"
-          :max="maxDegree"
-          :drag-on-click="true"
-          @change="onTemperatureChange"
-          @dragging="onTemperatureChange"
-        />
-        <period-select :selected-value="$store.state.selectedTemperatureType" :options="temperatureOptions" :value="viewTemperature" @change="onTemperatureTypeChange" />
-      </div>
-    </client-only>
-    <div class="slider-box-wrapper">
-      <div>
-        <div class="slider-box purple">
-          <img width="18" height="16" src="~/assets/icons/solid.svg" />
+    <div :class="{ grayscale: !active, 'blur-sm': !active, 'pointer-events-none': !active }" class="temp-container">
+      <client-only>
+        <div style="display: flex; margin-top: 8px" class="table-slider">
+          <vue-range-slider
+            ref="slider"
+            v-model="viewTemperature"
+            :tooltip="'none'"
+            :min="minDegree"
+            :max="maxDegree"
+            :drag-on-click="true"
+            @change="onTemperatureChange"
+            @dragging="onTemperatureChange"
+          />
+          <period-select
+            :selected-value="$store.state.unit"
+            :options="temperatureOptions"
+            :value="viewTemperature"
+            :min="minDegree"
+            :max="maxDegree"
+            @typeChange="onTemperatureTypeChange"
+            @valueChange="onTemperatureChange"
+          />
         </div>
-        <div class="title">Katı</div>
-      </div>
-      <div style="margin-left: 30px">
-        <div class="slider-box blue">
-          <img width="20" height="19" src="~/assets/icons/liquid.svg" />
-        </div>
-        <div class="title">Sıvı</div>
-      </div>
-      <div style="margin-left: 30px">
-        <div class="slider-box teal">
-          <img width="20" height="16" src="~/assets/icons/gas.svg" />
-        </div>
-        <div class="title">Gaz</div>
-      </div>
-      <div style="margin-left: 30px">
-        <div class="slider-box orange">
-          <img width="13" height="20" src="~/assets/icons/undefined.svg" />
-        </div>
-        <div class="title">Belirsiz</div>
-      </div>
+      </client-only>
+      <States />
     </div>
   </div>
 </template>
 
 <script>
 import 'vue-slider-component/theme/antd.css'
-import PeriodSelect from './PeriodSelect'
+
 export default {
   name: 'Slider',
-  components: { PeriodSelect },
   data() {
     return {
+      active: false,
       temperature: this.$store.state.temperature,
       viewTemperature: this.$store.state.temperature,
-      timer: null,
       tempTimer: null,
       temperatureOptions: [
         { label: '°C', value: 'c' },
@@ -72,18 +59,18 @@ export default {
   },
   computed: {
     minDegree() {
-      if (this.$store.state.selectedTemperatureType === 'c') {
+      if (this.$store.state.unit === 'c') {
         return -273
-      } else if (this.$store.state.selectedTemperatureType === 'k') {
+      } else if (this.$store.state.unit === 'k') {
         return 0
       } else {
         return -459
       }
     },
     maxDegree() {
-      if (this.$store.state.selectedTemperatureType === 'c') {
+      if (this.$store.state.unit === 'c') {
         return 7000
-      } else if (this.$store.state.selectedTemperatureType === 'k') {
+      } else if (this.$store.state.unit === 'k') {
         return 7273
       } else {
         return 12632
@@ -94,46 +81,48 @@ export default {
     '$store.state.temperature': {
       handler() {
         this.viewTemperature = this.$store.state.temperature
-        clearTimeout(this.timer)
-        this.timer = setTimeout(this.triggerTimer, 3000)
       },
     },
     temperature() {
       this.$store.commit('UPDATE_TEMPERATURE', this.temperature)
-      clearTimeout(this.timer)
-      this.timer = setTimeout(this.triggerTimer, 3000)
     },
-  },
-  beforeDestroy() {
-    clearTimeout(this.timer)
+    active() {
+      if (!this.active) {
+        this.$store.commit('DEACTIVATE_TEMPERATURE')
+      }
+    },
   },
   methods: {
     onTemperatureTypeChange(type) {
-      if (type === 'c' && this.$store.state.selectedTemperatureType === 'f') {
+      if (type === 'c' && this.$store.state.unit === 'f') {
         this.viewTemperature = Number(((this.viewTemperature - 32) / 1.8).toFixed(0))
       }
 
-      if (type === 'f' && this.$store.state.selectedTemperatureType === 'c') {
+      if (type === 'f' && this.$store.state.unit === 'c') {
         this.viewTemperature = Number((this.viewTemperature * 1.8 + 32).toFixed(0))
       }
 
-      if (type === 'c' && this.$store.state.selectedTemperatureType === 'k') {
+      if (type === 'c' && this.$store.state.unit === 'k') {
         this.viewTemperature -= 273
       }
 
-      if (type === 'k' && this.$store.state.selectedTemperatureType === 'c') {
+      if (type === 'k' && this.$store.state.unit === 'c') {
         this.viewTemperature += 273
       }
 
-      if (type === 'f' && this.$store.state.selectedTemperatureType === 'k') {
+      if (type === 'f' && this.$store.state.unit === 'k') {
         this.viewTemperature = Number(((this.viewTemperature - 273) * 1.8 + 32).toFixed(0))
       }
 
-      if (type === 'k' && this.$store.state.selectedTemperatureType === 'f') {
+      if (type === 'k' && this.$store.state.unit === 'f') {
         this.viewTemperature = Number(((this.viewTemperature - 273 - 32) / 1.8).toFixed(0))
       }
-      this.$store.commit('SET_TEMPERATURE_TYPE', type)
-      this.$store.commit('UPDATE_VIEW_TEMPERATURE', Number(this.viewTemperature))
+      clearTimeout(this.tempTimer)
+
+      this.tempTimer = setTimeout(() => {
+        this.$store.commit('SET_TEMPERATURE_TYPE', type)
+        this.$store.commit('UPDATE_VIEW_TEMPERATURE', Number(this.viewTemperature))
+      }, 150)
     },
     onTemperatureChange(val) {
       this.viewTemperature = val
@@ -143,9 +132,6 @@ export default {
     triggerSetTemperature(val) {
       this.temperature = val
     },
-    triggerTimer() {
-      this.$store.commit('DEACTIVATE_TEMPERATURE')
-    },
   },
 }
 </script>
@@ -153,36 +139,35 @@ export default {
 <style lang="scss">
 @import '~@/assets/css/partials/variables';
 
+.slider {
+  grid-area: 1 / 3 / span 3 / span 10;
+  padding: 30px;
+}
+
 .vue-slider {
   padding: 0 !important;
 }
 
 .table-slider {
   .vue-slider {
-    width: 30vw !important;
-  }
-}
-
-.element-slider {
-  .vue-slider {
-    width: 30vw !important;
+    width: 100% !important;
   }
 }
 
 .vue-slider-rail {
   background-color: #0b0e13 !important;
   border-radius: 5px !important;
-  height: 8px !important;
+  height: 10px !important;
 }
 
 .vue-slider-dot {
-  background-color: #66ccca !important;
+  background-color: #e8bc0b !important;
   border-radius: 50%;
 }
 
 .vue-slider-process,
 .vue-slider-dot-handle {
-  background-color: #66ccca !important;
+  background-color: #e8bc0b !important;
 }
 .vue-slider-dot-handle {
   border: 1px;
@@ -190,44 +175,40 @@ export default {
 }
 
 .slider-title {
-  font-size: 10px;
+  font-size: 16px;
+  font-weight: 600;
   color: $gray;
 }
 
-.slider-box-wrapper {
-  display: flex;
-  margin-top: 23px;
-  text-align: center;
-  .slider-box {
-    width: 2.3vw;
-    height: 2.3vw;
-    border-radius: 4px;
-    display: flex;
-    img {
-      align-self: center;
-      margin: auto;
+.temp-container,
+.temp-container * :not(.vue-slider-rail *) {
+  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
+}
+
+.temp-container.pointer-events-none * {
+  pointer-events: none !important;
+}
+
+.vue-switcher-theme--custom {
+  &.vue-switcher-color--blue {
+    div {
+      background-color: lighten(#66ccca, 10%);
+
+      &::after {
+        // for the circle on the switch
+        background-color: darken(#66ccca, 5%);
+      }
     }
-  }
-  .purple {
-    box-shadow: 0 2px 40px 0 rgba(185, 148, 250, 0.4);
-    background-image: linear-gradient(135deg, #d5b7ff, #aa80ff);
-  }
-  .blue {
-    box-shadow: 0 2px 20px 0 rgba(136, 158, 255, 0.4);
-    background-image: linear-gradient(135deg, #acbdff, #8095ff);
-  }
-  .teal {
-    box-shadow: 0 2px 20px 0 rgba(153, 255, 252, 0.4);
-    background-image: linear-gradient(135deg, #80fffc, #56d9d7);
-  }
-  .orange {
-    box-shadow: 0 2px 40px 0 rgba(255, 175, 128, 0.4);
-    background-image: linear-gradient(135deg, #ffaf80, #ed954b 100%);
-  }
-  .title {
-    font-size: 10px;
-    color: $gray;
-    margin-top: 3px;
+
+    &.vue-switcher--unchecked {
+      div {
+        background-color: lighten(#66ccca, 30%);
+
+        &::after {
+          background-color: lighten(#66ccca, 10%);
+        }
+      }
+    }
   }
 }
 </style>
