+++ package.json
@@ -26,6 +26,7 @@
   },
   "dependencies": {
     "@nuxtjs/axios": "^5.13.1",
+    "@shopify/draggable": "^1.0.0-beta.8",
     "axios": "^0.21.1",
     "body-parser": "^1.19.0",
     "connect-timeout": "^1.9.0",
@@ -35,7 +36,7 @@
     "dotenv": "^8.2.0",
     "element-ui": "^2.15.1",
     "express": "^4.17.1",
-    "fuse.js": "^6.4.6",
+    "fuse.js": "^6.5.3",
     "helmet": "^4.4.1",
     "install": "^0.13.0",
     "lodash": "^4.17.21",
@@ -53,7 +54,9 @@
     "v-tooltip": "^2.1.3",
     "vue": "^2.6.12",
     "vue-awesome-swiper": "^4.1.1",
+    "vue-fuse": "^2.1.1",
     "vue-gtag": "^1.16.1",
+    "vue-js-toggle-button": "^1.3.3",
     "vue-sanitize": "^0.2.1",
     "vue-slider-component": "^3.2.11",
     "xlsx": "^0.16.9"
