+++ plugins/vue-fuse.js
@@ -1,0 +1,4 @@
+import Vue from 'vue'
+import VueFuse from 'vue-fuse'
+
+Vue.use(VueFuse)
