+++ plugins/vue-toggle-button.js
@@ -1,0 +1,4 @@
+import Vue from 'vue'
+import ToggleButton from 'vue-js-toggle-button'
+
+Vue.use(ToggleButton)
