+++ store/state.js
@@ -1,16 +1,22 @@
 export default () => ({
   app: {},
   elements: [],
+  reducedElements: [],
+  searchResults: [],
+  selectedElement: null,
+  fuse: null,
   compounds: [],
   isotopes: [],
   probableElements: [],
   temperature: 0,
-  selectedTemperatureType: 'c',
+  unit: 'c',
   isTemperatureTriggered: false,
   isIsotopeFetched: false,
   isCompoundFetched: false,
   showInfoModal: false,
   selectedContentId: null,
+  selectedCategory: null,
+  selectedBlock: null,
   searchText: null,
   isMobile: false,
   isOriented: false,
